Move QueueConfig into ert.config
