+++ src/ert/_c_wrappers/enkf/__init__.py
@@ -8,7 +8,6 @@ from .ert_config import ErtConfig
 from .ert_run_context import RunContext
 from .model_config import ModelConfig
 from .observations import GenObservation, ObsVector, SummaryObservation
-from .queue_config import QueueConfig
 from .row_scaling import RowScaling
 from .run_arg import RunArg
 from .time_map import TimeMap
@@ -23,7 +22,6 @@ __all__ = [
     "HookRuntime",
     "AnalysisIterConfig",
     "AnalysisConfig",
-    "QueueConfig",
     "EnsembleConfig",
     "EnkfObs",
     "ModelConfig",
