+++ src/ert/config/__init__.py
@@ -2,6 +2,7 @@ from .ext_param_config import ExtParamConfig
 from .field_config import Field
 from .gen_data_config import GenDataConfig
 from .gen_kw_config import GenKwConfig
+from .queue_config import QueueConfig
 from .summary_config import SummaryConfig
 from .surface_config import SurfaceConfig
 
@@ -12,4 +13,5 @@ __all__ = [
     "ExtParamConfig",
     "SurfaceConfig",
     "Field",
+    "QueueConfig",
 ]
