+++ src/ert/ensemble_evaluator/_builder/_ensemble_builder.py
@@ -10,7 +10,8 @@ from ._legacy import LegacyEnsemble
 from ._realization import RealizationBuilder
 
 if TYPE_CHECKING:
-    from ert._c_wrappers.enkf import AnalysisConfig, QueueConfig
+    from ert._c_wrappers.enkf import AnalysisConfig
+    from ert.config import QueueConfig
 
     from ._ensemble import Ensemble
