+++ src/ert/ensemble_evaluator/_builder/_legacy.py
@@ -26,8 +26,9 @@ from .._wait_for_evaluator import wait_for_evaluator
 from ._ensemble import Ensemble
 
 if TYPE_CHECKING:
-    from ert._c_wrappers.enkf import AnalysisConfig, EnsembleConfig, QueueConfig, RunArg
+    from ert._c_wrappers.enkf import AnalysisConfig, EnsembleConfig, RunArg
     from ert.callbacks import Callback
+    from ert.config import QueueConfig
 
     from ..config import EvaluatorServerConfig
     from ._realization import Realization
