+++ src/ert/run_models/base_run_model.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 import concurrent
 import logging
@@ -8,12 +10,12 @@ from abc import abstractmethod
 from contextlib import contextmanager
 from functools import singledispatchmethod
 from pathlib import Path
-from typing import Any, Dict, Iterator, List, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Union
 
 from cloudevents.http import CloudEvent
 
 import _ert_com_protocol
-from ert._c_wrappers.enkf import EnKFMain, QueueConfig
+from ert._c_wrappers.enkf import EnKFMain
 from ert._c_wrappers.enkf.enums import HookRuntime
 from ert._c_wrappers.enkf.ert_run_context import RunContext
 from ert.callbacks import forward_model_exit, forward_model_ok
@@ -33,6 +35,9 @@ from ert.storage import StorageAccessor
 
 event_logger = logging.getLogger("ert.event_log")
 experiment_logger = logging.getLogger("ert.experiment_server.base_run_model")
+
+if TYPE_CHECKING:
+    from ert.config import QueueConfig
 
 
 class ErtRunError(Exception):
